Move nhentai models out of nhentaiScraper.cs
